add grpc LegislationClient

+++ PaySmartly.Calculations/Legislation/InMemoryLegislation.cs
@@ -13,7 +13,7 @@ namespace PaySmartly.Calculations.Legislation
             TaxableRange range2 = new(14_001, 48_000, 0.175);
             TaxableRange range3 = new(48_001, 70_000, 0.3);
             TaxableRange range4 = new(70_001, 180_000, 0.33);
-            TaxableRange range5 = new(180_001, int.MaxValue, 0.39);
+            TaxableRange range5 = new(180_001, double.MaxValue, 0.39);
 
             ReadOnlyCollection<TaxableRange> ranges = new([range1, range2, range3, range4, range5]);
             taxableIncomeTable = new(ranges);
